Remove getConfigurationLocation from IConfigurationStore
Why:
IConfigurationStore is supposed to be a generic interface
for storing data. Location information is needed for a lot
of (most of?) kind of stores, but:

    (1) definitely not for all (StaticStore doesn't need any location)
    (2) it means something completely different in each implementation,
        making it nearly impossible to document a shared meaning for
	IConfigurationStore::getConfigurationLocation().

Instead, let's use instanceof and methods within specific implementations
to work with configuration data. That way, we don't need a common signature
or a common method. WikiPageStore, our main store, can return a Title,
whether a possible future DB based store can return something else,
like a DB connection and/or a table name.

The only place where IConfigurationStore::getConfigurationLocation()
is used is ValidationHooks, and it can be easily avoided there.

This also fixes a TODO in ValidationHooks, since the JsonValidate
hook now only runs where it really should, not also when location
data for a non-wikipage provider accidentally collides with a wikipage's name.

What:
* Remove IConfigurationStore::getConfigurationLocation()
* Rewrite ValidationHooks to not rely on location in the store