Editor: send new config data on form submision
- Use action=communityconfigurationedit API to submit
new configuration data when user clicks on the form submit
button.
- Make the configuration state (jsonform.data) a Vue reactive
object rather than a reference to the initial configData passed
to the form. This is to avoid unconventionally writing into a
prop (props.data) and to keep a copy of the original data in
case is needed for diffing.
- Make the UI schema aware of required top-level properties
so rendered inputs trigger browser HTMLInputElement validation

Bug: T356622
Change-Id: Ic25a184e68747dbfd10a2f6c2de1af4ad8f13677