Editor: add PageTitleControl component
- Apply the PageTitleControl to data schemas
that use a reference to the type definition
'PageTitle' as its type
- The control requests page titles thorugh the
action api prefixsearch that match a given search
term and display them as menu items.
- Use the page title string as the model
for the given property
- To be used with the newcomer homepage schema
example in https://gitlab.wikimedia.org/repos/growth/community-configuration-example/-/merge_requests/9

Possible follow-ups:
- Allow to use pageid number or a page object
as the model for this control
- Display an extra menu item at the top matching
the currentSearchTerm to allow to input a non
existent page. Similar to HTMLTitleTextField
behavior.
- Load more items on scroll is possible although
not sure how useful it is for our use cases.

Bug: T357712
Change-Id: Idb707870f2a09439e0ee479761ddd0eb4c3d2c3f