Use mixed type for config objects
Why: using array type is restrictive to
other eventual top-level type schemas and
makes hard to distinguish between JSON
arrays and JSON objects.

What: move to a mixed type for now. This
will be stdClass with current status in which
top-level schemas are hardcoded to be objects
but gives some flexibility to move to more
restrictive approach if necessary

Note: when no revision of a config page
is found stdClass is returned. If
different type from object is used for
the top level schema, a different empty
default will need to be provided.

Bug: T360148
Change-Id: Id5e21f643c096800aa0c9f93fa1766724c1a3397