Respect JsonSchema defaults
Why:
The schema is supposed to completely describe how the configuration
page should work like. Since it has support for defaults, it only
makes sense to use JSONSchema-provided defaults rather than falling
back to GlobalVarConfig (as we used to do in CC1.0).

This patch reserves a default of `null` to represent "no default at all".
This is because core's ReflectionSchemaSource uses null as the super-default
(aka default to use when the user did not fill any), and it only makes sense
to interpret "default key is not defined" as "no default is used".

If we find ourselves in need of `null` as an actual default, we will likely
need to revisit this decision.

In addition to that, this patch only support default key at the top level.
Nested objects are expected to have their default defined as one big blob.

What:
* Add support for JSON Schema defaults
* Fix PHPDoc @return for SchemaBuilder::getRootProperties() (Phan complained
  about it being inconsistent with the code).

Bug: T360042
Change-Id: Ida2e20274b35c5de3b55bc37eeb1f24108eb3adc