WikiPageConfigProvider: Fallback when store returns invalid data
Ocasionally, the store might return invalid configuration,
or might be unavailable for other reasons. When that happens,
we should fallback to the default config, to ensure the site
behaves at least somehow reasonably.

With T360042 being resolved in the parent patch, we can easily
get the defaults to fallback to using SchemaBuilder::getDefaultsMap().

Bug: T359124
Bug: T351517
Change-Id: Id80a9a9103391e6ea736558b7b074b5e8d8d921b