Display meaningful error message when config fails to load
Why:
To ensure admins have a chance of fixing accidental validation
errors, provide them with two key pieces of information:

    (a) A detailed validation error message
    (b) A link to the on-wiki JSON page (where validation
        error can be fixed).

To maintain a sufficnetly generic approach, introduce a concept
of info links on the store level. That way, the store (as the only
piece of CommunityConfiguration which is title-aware) can inform
the UI where the user should be directed.

When a provider that's not based on on-wiki JSON page is introduced,
the info link can be changed to a documentation page (or omitted
altogether). This is also why the i18n message directs the user
to the info link in a generic way (using "see"), avoiding any promises
about the link's content.

What:
* Introduce IConfigurationStore::getInfoPageLinkTarget()
* For the WikiPageStore, set the info page link to the JSON page
  itself. For other providers, it will likely direct to appropriate
  documentation.
* Expose the validation error to the user in GenericFormEditorCapability.
  As of now, it will be English-only (T351879 exists to change that).

Bug: T360368
Change-Id: I9b75a20902f48ce52bdd62fa32d30bc0eed7aced