Refactor permission checks
This moves permission handling from Writer to the WikiPageStore,
making Writer a helper-only class that changes wikipages when requested,
but otherwise does not interfere with the process.

The patch also properly documents what method in IConfigurationStore
is and is not expected to run permission checks. Most importantly, it
removes a comment about missing permission checks from the DataProvider.

The changes are made, as I today confused myself about missing
permission checks (relying on the docstrings without checking
whether they are consistent with the reality).

Change-Id: Ifc027e400c772423566fa5f2968c477848263e78