Expose older schema versions
This patch introduces a concept of versions
into CommunityConfiguration. The schema version
a given configuration was created with is stored
within the store, using a reserved field called $version.

The version data is stored by the store. Depending
on the implementation, it might be stored alongside
the data (as in WikiPageStore), or in a separate location.

Version data is appended shortly before saving and removed
before returning data to the caller. If version data is
required, it needs to be requested separately.

Bumping a schema version means copying the old schema to
a Migration subnamespace and bumping the VERSION constant
in the class. This should be automated by a script in the
future.

The patch provides foundation for migration logic to be able
to access the current version, as well as any previous version
of a schema.

Bug: T357532
Change-Id: I8e1f2ef2e403ba28eb4d66eb938128a79614fcdb