Factor AbstractProvider out of DataProvider
The idea around DataProvider was to have a basic
provider which returns the complete content of the store
(while guaranteeing it is valid), and does not provide
any additional services.

However, it evolved into a base class for all providers,
containing implementation even for getName() or getStore(),
which is something that should really be common for all providers,
including ones that end up not being DataProvider for some reason.

Hence, this patch moves the common code from DataProvider into
an AbstractProvider.

This will be useful in I8e1f2ef2e403ba28eb4d66eb938128a79614fcdb,
which adds AbstractMethod::saveConfiguration (which does not belong
into a data provider conceptually).

Change-Id: I73c7d0ba4df6faf1120fb2cdb71cac460d5b5c41