Show the provider name in the editor title, not the Id
The provider name is an actually localized title, instead of a sometimes
cryptic internal identifier.
As the key of a message usually does not contain uppercase characters,
this adjust the `AbstractProvider::getName` method to also always
lowercase the provider id. This is aligned with the behavior in
DashboardEditorCapability::getProviders(). It is a bit unfortunate that
this logic to generate the message key is duplicated in two places.

The message returned by `getName` is explicitly escaped to make it
obvious that it is secure, even though `setPageTitle` explicitly also
does escaping of bad tags per its contract.

The tests are adjusted accordingly, for consistency's sake, all the
provider IDs in the test file have been changed to a string that
includes capital letters.

Bug: T361933
Change-Id: If8cd7f7311b161a0fa65d31158bbeb49e2ea8659