Rename parameter and private field providerName to providerId
This *is* the id of the provider and also what is returned by getId().
Renaming it should reduce confusion when looking at the class and
instantiating it.

This is a pure name change of a private field and a constructor
parameter and so should not have any effects outside whatsoever.

See also T361931 for a more comprehensive treatment of this issue.

Bug: T361933
Change-Id: If8df085e26492468eebbde209e2ded280bf7a983