Add test for SpecialCommunityConfiguration
This was originally intended as a test for the previous change that
modifies the page title, until I noticed that the page title is not
actually available.

Still, it is a good opportunity to cover the special page with tests.

Also, this removes a spurious space from one of the templates that was
annoying in the tests.

Change-Id: I9011d6d436f875b0d41b29350c9a8d7d596a1b09