Introduce AbstractProvider::storeConfiguration
Having the provider able to change the config before
being saved is useful for a couple of reasons.

For example, in T357532, we want to keep the $version
parameter within the page, so that we can know what
schema version the file was created against.

In Help panel, we need to autocompute a property,
which is best suited at the provider level, but
there is currently no place where such manipulation
could happen.

To resolve those needs, add AbstractProvider::storeConfiguration,
which could perform such manipulations. Right now, none
is included.

This used to be within I8e1f2ef2e403ba28eb4d66eb938128a79614fcdb,
but since that requires some finishing up, factoring it
to a separate patch.

Bug: T357532
Bug: T360472
Change-Id: Iab1c0c107a6f5a54d51ccf32fb766a33b7c86327