Editor: add control for array type
- Modify the uischema/i18n generators to use
the configuration data when building the array
control properties and add a "labels" array.
It has the label (control-label and help-text
are left out for now) for each array element field.
- Add a Vue control component ArrayControl which
iterates through existing elements in the configuration
and dispatches appropriate renderer.
- Add useJsonFormArrayControl composable to extend the
array control bindings with a childUISchema property which
will hold the uischema for the array items.
- Update the state getter/setter functions to
correctly parse number-indexed array property scopes
(similar to a json pointer), eg: "#/properties/someArrayProp/0"

Bug: T360472
Change-Id: I600347aec0baa403e2c96e02dc2d897954df2005