Add validation warnings
Only warning about some issues instead of failing will be necessary when
migrating a Schema to a newer version that might have new parameters or
might drop some.

This could in principle be implemented with the existing
`IValidator::validate` method by fully relying on the difference between
`StatusValue::isOK` and `StatusValue::isGood`. However, that difference
is subtle and prone to confusion and thus errors. Though this could be
mitigated with thorough and explicit testing.
For now, this change introduces a second method on `IValidator` to make
it explicit that this will return a StatusValue that is still `OK` if
there are warnings. `::validate` has been renamed to make the intended
functionality explicit.

For now, this change assumes that IValidator will not be stable neither
to implement nor to call by external code. If this assumption changes,
then we may want to reconsider whether these public method signatures
are the best to make stable, or whether we should find a solution that
is more extendable under a stable interface policy. This could, for
example, mean to only expose a single ::validate() method that then
would have a second "mode" parameter that could either be
`IValidator::FOR_READS` or `IValidator::FOR_WRITES`.

Bug: T362042
Change-Id: Ie53284c56f43684256bb3a8cb0e10cf4fd24986e