PageTitlesControl: close menu when there is no input
Fix for the menu remaining visible when there is no
user input. That happens when the user types some
text and then completely deletes it.

Also close the menu when the field looses its
focus and open it when it gains it.

Bug: T357710
Follow-up: I384eb017f41e665a43583aac2620157d94e61ffd
Change-Id: I04d018a96978a9b96c7c512d7f32731b597925f0