Always set `is-fieldset="true"` for checkbox fields per codex docs
The codex docs specify that checkboxes in a form field must always set
`is-fieldset="true"` on the field. Even if it is just a single checkbox.
This is apparently necessary for some accessibility optimizations. The
fieldset is not visually rendered.

See: https://doc.wikimedia.org/codex/latest/components/demos/checkbox.html#form-field

Change-Id: I852a6f77754c3f3a7a10721f21d6eef56ac9e949