Add tests for GenericFormEditorCapability
This also move the instructions for how to instantiate the
GenericFormEditorCapability class with ObjectFactory into said class.
This has two advantages: those instructions live in the same place as
the constructor with which they need to align, and they can be used in a
test as they are now no longer private.

Change-Id: I6cf6b12070d830094d789fd75dce4d2e1f3e6e4e