BooleanControl: unwrap modelValue, drop dead code, add tests
Values returned by setup are shallow unwrapped. However, since this
component is using a field of an object as the modelValue, we have to
unwrap it ourselves.
See https://vuejs.org/api/composition-api-setup.html
This takes care of the vue warning:
`[Vue warn]: Invalid prop: type check failed for prop "modelValue".
Expected Bool | Array, got Object`
However, now there is the following warning on toggle:
`Write operation failed: computed value is readonly`
That is because currently in `useJsonFormControl` a computed property is
used, which is read-only. This needs to be replaced by a `ref`, but that
is beyond the scope of this change.

Further, the prop `controlLabel` was unused, because the label from the
slotProps, that is, from the uischema is used.

Also, this now adds a test which asserts that the control label is indeed used
for the input and that toggling the input does mutate the data.

Change-Id: I0d57e0a499360c8642fc50769935b32a72d3fa25