Add support for enums
This only adds the message keys for the radio input labels to the
uischema instead of the message objects as a first step to move away
from using the global `mw.Message` inside lib/json-form.

This has the side-effect that the labels will show as their message
keys if they are missing. However, since they are strictly required in
order for that control to make sense, this seems acceptable.

For now, enums can be fundamentally of string or number type. Other
types can be added later if we consider them needed and sensible.

The renderer gets a higher rank than the basic controls, because in
addition to checking the type attribute, it also checks for the presence
of the enum attribute. Thus it must take explicit precedence of the
isStringControl and isNumberControl ones.

Bug: T362685
Change-Id: Ia58213956daea1972415d6b41a235ef478a18fe4