Editor navigation: add history and edit tab links
The tabs allow users to easily access the page history and
edit functions. Append a "View form" tab and sort the `views` links
so it displays as the first navigation tab in right top corner.

Also remove the rest of not used links by design, see designs
in the task description.

Bug: T363788
Change-Id: I3be36874480e9c6dc9a08f0b074910127b2e409f