Do not let callers to modify Loader::load's internal cache
When used with a stdClass (or another object), StatusValue::getValue()
will return a reference of the value, not a copy. This means
that before this patch, manipulating with the value
after receiving it from Loader::load() was being propagated
to the in-process cache, resulting in bugs that cost 4+ hours
to debug.

The only way to get a StatusValue from Loader
is by calling the public entrypoint (Loader::load).
This means we can fix the problem by doing
the deserialization in Loader::load rather than
in the fetchConfig() private method.

Callers of Loader::load() are told to not modify the returned
StatusValue itself, as only the data is guaranteed
to be mutable.

Special thanks to Roan for helping with the solution
at the 2024 Hackathon.

Co-authored-by: Roan Kattouw <roan.kattouw@gmail.com>
Bug: T364101
Change-Id: Ie0673f33a9cc8ed2b7582e2aa3f6b703eee7d2e2