Complete the tests for for i18n js
This will be helpful to show what changes in the following
refactor/bug-fix change.

While the test for i18n.getControlTextProps() does not assert the
_structure_ of the returned message, it does assert that all the needed
keys are there.

Note that regarding T364527, this asserts the presence of more keys than
needed. This is done, so that the following change fixing that bug can
be easier reviewed by only having to review the _changes_ to the tests
added here, and not the tests themselves.

Change-Id: I0049e13baf1c20d2c2dd37bd2c9293a7c6a717c1