Move caching logic to AbstractJsonStore
This will make it possible to introduce additional store
types with ease.

Originally uploaded with a HttpStore in mind (cf. T359185),
but it simplifies the logic both for WikiPageStore
and any possible future store that'd be making use
of caching, so I think it makes sense to merge it
either way.

Change-Id: Ia5699e2cea1b4a36228ed713cdedac2ab71019b0