More fine-grained handling of i18n for custom controls
This now no longer uses the getStringControlMessages function
in i18n.js for PageTitle controls, but its own dedicated function.

This also allows us to load the no-results message the same way as all
the other messages and only when the pagetitle control is actually used.

Separating the PageTitle control on the one hand, and the PageTitles and
Namespaces controls on the other hand will be also useful when
hard-coding a placeholder for the latter, but allowing the custom
definition of one for the former.

Change-Id: I730b7e7b17a22430c472bd97cc31df070a103802