[tests] Update covers annotation
== Remove per-method @covers
According to PHPUnit documentation,
it is not recommended to use per-method @covers
annotation, and they are also not supported
at all in the attributes annotation (we do not
use that yet, but we will likely migrate to that
at some point).

== Explicitly cover abstract classes
Since it is impossible to test abstract classes
alone, include them in @covers annotation
for their children.

Change-Id: Ic7c759b83246e6bdd21557cd230c9d37dac0c83f