Editor: fix invalid prop warnings
Codex CdxTextInput expects v-model to be of
types string|number rather than a Vue computed
object.

Passing the inner value revelaed that we used
`modelValue = computed( value )` in useJsonFormControl
which created computed readonly refs which are not
suitable for `v-model` binding which expects a writable
ref. Vue did not complain about it becauuse we use a custom
onChange handler for the model updates. But it did after
passing the inner value. Change computed by ref in
useJsonFromControl.

Bug: T365240
Change-Id: Ic03d7044fd6999633ddabd0f86087109bd211307