JsonSchemaBuilder: Accept arrays as defaults for TYPE_OBJECT
Due to PHP language limitations, it is not possible
to use PHP objects in constants (considering objects
are always mutable). Since we have schemas represented
as PHP classes with constants, that poses a problem
for using JsonSchema::DEFAULT with TYPE_OBJECT.

To avoid this problem, we make JsonSchemaBuilder
automatically convert arrays to objects when TYPE_OBJECT
is encountered. That way, defaults of objects can still
be used in a easy-to-understand way.

This is intentionally not accounting for "objects in objects",
or "objects in arrays". For that, we might want to reconsider
the "defaults only at the top level" approach, and instead
use something like JsonSchemaTrait::getDefaultFromJsonSchema
is doing for objects in objects. Since we do not need
objects in objects now, this is left as a future improvement,
to give enough time for the discussion.

Bug: T365045
Change-Id: I32f4318f7bcd3d42151423d87882ecded48d2f72