Writer: avoid storing empty arrays and order config keys
When $newConfig is an empty object, an empty array
is stored due to the (incorrect) sorting processing. Avoid
it by explicitly casting $newConfig to object before encoding.

uasort is called with the intent of sorting config keys
alphabetically but failing to do so because it compares
values rather than keys in an associative array. Use
ksort instead.

Bug: T366168
Change-Id: I1ec7d9f9b67c47955e94ad7f9bec5576bc6f9c8e