Rework Editor messages system and improve validation error message
This change does two things:
1. It reworks the way we display messages in the app
2. It improves the display of validation messages

How we display messages in the app
==================================

The previous system had grown very unwieldy: it was controlled by 5
different, partially optional refs, and its template consisted if
several if-sections that were partially nested and in a level of detail
that was not really related to the App component.

This system now allows for dedicated Message components. That allows us
to evolve messages individually without having to make one single
generic template handle more and more special cases.

This also has the large advantage that the template is now much shorter,
and so easier to move to a fixed footer, when that will be created for
T364500.

Improved Validation errors
==========================

The validation errors messages now makes use of the additional data
introduced in the previous change I5f94ea9aeae15212116bfc458e89d15245371370
to also link to the field that has the validation problem. A follow-up
change will add inline messages to the specific fields explaining what
is wrong there: Ia93c6948889ef510af59d0d276b3dadeba5f9cb1

Bug: T363855
Change-Id: Ie0c49298a587e83c292404d5c69eced119ae7f9e