Drop logging level for unsupported providers to DEBUG
WikiPageConfigReader iterates through all providers and
as such, it is perfectly reasonable for it to encounter
a provider that is not a WikiPageConfigProvider. Lower
the loglevel to debug, as this is a perfectly legitimate
situation for the app to be in.

Bug: T366519
Bug: T360954
Change-Id: I7ffd5a65af168bec463f85a401f1963617c8a5bd
(cherry picked from commit 4b96db51f06e1438f3abb98eeef803bc8791da36)