Fix warnings in jest tests
This removes defineComponent() from EditSummaryDialog.vue as this led to
the warning " Component is missing template or render function."

Also, it adds a few missing mocks and directives that caused warnings.
They might need a bit more functionality once that is being asserted.

Finally, it asserts that console.warn and console.error are not called
unexpectedly. Doing it in the same way as GrowthExperiments.

Change-Id: I2d58ca2b814a1e91a6664e2e13c2582ac69d58b1