Add array management
The `forceRerenderCounter` is necessary to ensure the list is correctly
updated when an item is removed. This is because we are using the item's
index in the key, which is an anti-pattern. Ideally, we would use
something that is specific to each respective array item there, but that
is not possible, because we have no "id" field there.

See also: https://vuejs.org/guide/essentials/list#maintaining-state-with-key

Adding and using data-test-id attribute is a technique from Cypress that
is also sometimes useful in jest to make tests less brittle by not
binding to things the might change like classes, ids, copy.
See also: https://docs.cypress.io/guides/references/best-practices#Selecting-Elements

Bug: T366274
Change-Id: I8a83adb16258708d2da64159cbe1bdf3994fcde1