Add hook for dynamically hiding form fields
Register a new hook to be used for modifying
a schema before it is passed over to the frontend
for form rendering. This is intended for dynamically
hiding form fields based on wiki's state.

The patch is based on Michael Große's work
in I1f191129308bedd9d9bd045d987db8db653d30aa.

Bug: T367331
Change-Id: Ia0282d136bff06aa4306027aa7f30fbb4c7eb6f0