WikiPageWriter: Do not use the clone keyword
WikiPageWriter clones $newConfig, and immediately
casts the cloned data into an array. This is unnecessary,
because arrays are always passed by value, so the clone
does not do anything useful. We have a test
(added in Ie94d01fd3c83bee2e20cc14c4df104a703162596) to
ensure Writer does not write back to $newConfig, and since it
passes, this conclusion is confirmed.

It also means WikiPageStore::storeConfiguration() does not,
in fact, accept "any JSON serializable type" (as documented
in its PHP docstring). Since the clone keyword is unnecessary,
remove it and meet the documented feature. Add a testcase
to WikiPageStoreIntegrationTest to verify the requirement is met.

Bug: T368741
Bug: T367575
Change-Id: I3c07a3060ba9cda5ff9304b31262fa6297188804