Allow client extensions to provide their own tag names
Since stores and their construction needs to be independent
on the storing backend, we cannot add tags as part of
the method signatures. Instead, the patch adds a concept
of store options, which can be used to modify the store's behavior.

In this case, we use store options to add extra tags to individual
edits, and use it to honor the "mentor list change" tag GrowthExperiments
used before.

Bug: T368739
Change-Id: Iabb18a1038b94c430bb741f8c6d00a10ada303bc