Make it possible to skip permission checks from the provider level
Ifcbfda68df17f476b46d08b324bcaccf3fdfbdf8 added support
for bypassing permission checks from the store. While
that works, it is hard to use from outside of CommunityConfiguration,
as none of the stuff providers implement will be used by default,
most importantly, calling the validator.

To avoid that, this method adds IConfigurationProvider::alwaysStoreValidConfiguration,
which will instruct the store to bypass any permission checks.

To ensure validation is always performed, regardless of whether
permission checks are performed, this patch makes
AbstractProvider::doStoreValidConfiguration (previously named just storeConfiguration)
the center of storing. That is the only place where permissions mode is expressed
as a boolean, rather than a separate method, as it serves logic that needs to be applied
regardless of whether we care about permissions.

storeConfiguration is renamed to doStoreValidConfiguration to better express
its meaning, especially since validation is moved to this method.
Even if validation was kept in DataProvider (where it previously was),
it would still be in a function that does not have "valid" in its
name, which appears to be contrary to the naming pattern CommunityConfiguration
adopted.

Bug: T367575
Change-Id: I9c0b9f6adb4a805287e2fc547359c6e346c569b8