fix(editor): make PageTitleControl reliably blankable
This fixes the bug that blanking a pre-existing value and then saving
would result in a successful save without the value actually changing.

The primary cause for this is that LookupControl is not firing a
update:selected with `null` unless it previously had a "selected menu
item", see:
https://github.com/wikimedia/design-codex/blob/403c3b6b223759c8b74eac07b0b3544a9eef4794/packages/codex/src/components/lookup/Lookup.vue#L240-L246
This is fixed with two changes:
- setting the initial value also for the v-model `selected` ref
- doing an initial search for menu items via onInput

There is a second issue that unselecting an item makes CdxLookup emit
`null`, but we do not support that yet. So in onSelectionChange we
change that to `''` instead. That might be worth reconsidering once we
took care of T365145.

Bug: T370326
Change-Id: Id63a4d6472a269b4984225fe5094ba3e455d6718
(cherry picked from commit b2568100f779266ef3717ea8cbbe9c6ed2abc97e)