fix(migrateSchema): use correct converter when downgrading
It is the converter of the current version that knows what needs to
change for a config to take the shape of the Schema at the previous
version.

Note that this does not address the fundamental problem of T371028, but
it does fix the issue of there being an exception when trying to
downgrade to the first version.

Bug: T371028
Change-Id: I289d89d71682aa81c680d9755135e671d0019104