refactor(ISchemaConverter): rename downgrading method
`downgradeFromNewer` is misleading. The converter has no idea what needs
to change for a config from a Schema of newer version to be valid
against the Schema of its own version. What the converter does know, is
what needs to change from a config valid against its current version to
a config valid against the previous version.

Bug: T371028
Depends-On: I788714abdde9f515da5362e2d01c946e5687e732
Change-Id: I7debc14017771e5f183b0dc9695d0ba619822dd7