Hide providers with `excludeFromUI` flag from UI components
If a provider is excluded from the dashboard, we should
not display it in any portion of CommunityConfiguration-generated UI.
Previously, the `skipDashboardListing` flag would only
remove the provider from the dashboard listing, but
links like `view form` would be visible in other areas
like the JSON pages, which shouldn't be the case.

The `skipDashboardListing` flag has also been renamed
to `excludeFromUI` as it now controls visibility
across the entire UI, not just the dashboard.

Bug: T368728
Depends-On: Ifeb883825c9959bfc9b50f378230179c55273679
Change-Id: If1aee37c8cbbfcc4aaa070bf92af8cf121f1eebf