feat(SchemaBuilder): array-based `required` for top-level properties
This defines a special constant that holds the keys of the top-level
properties that are required. Fundamentally, this close to what the
DRAFT 4 (=current) defines.

This is currently an implementation that lives entirely within
CommunityConfiguration. But the details of the special handling could be
upstreamed to ReflectionSchemaSource, either the special handling for
that particular key, or for keys starting with `__` in general.

The overall alternative to this approach would be to read the
`SELF::REQUIRED => true` field from the top-level properties directly
and assemble the needed array from that.

`JsonSchema` itself does not have a `protected const __REQUIRED`,
because it would seem pointless. The only place where that could
possibly be used would be JsonSchemaReader::getRequiredTopLevelProperties
and that can't access it because it is not public and so it has to
define its fallback as a literal value.

See also: I03800389be48d23f3b55a0c0d22ca83b5d74a28d

Bug: T360921
Change-Id: I38861144a60a21432c8775552179eba1e1fa25f0