chore(eslint): unbreak LibUp by manually updating eslint-config-wikimedia
This somehow confused LibUp by causing a line-too-long eslint error
which in turn somehow caused LibUp to try to lint *all* files, including
json, with the vue/max-length rule, which in turn crashed because of too
many errors. For the most recent log with this error see:
https://libraryupgrader2.wmcloud.org/logs2/1442470

After this manual update, LibUp should hopefully work on
its own again.

Change-Id: Ifc85beedf4d6d06d331bd97ef4b7228636711d08