docs: Add minimal example implementation
This code will never run in production and so it should add as little
overhead as possible. Thus the configuration is not in the main
extension.json, where it would be parsed in production and hook handlers
run, etc., but in a separate example-extension.json which needs to be
actively added to LocalSettings.php or InitialiseSettings-labs.php for
beta.

Bug: T372028
Change-Id: Ie61ac51d58bbd0776d305b847acf1bf0b0203078