feat(editor): add global error handle to log errors to logstash
This logs the errors in a global error handler, because this can also
capture errors that occur before the error handler in App.vue is
initialized.

The fields use snake_case instead of camelCase because that is the
pattern used in the upstream schema by WikimediaEvents.

See I0226f664af8fa223e3d151b7e1fe8873a4781ae7 for error_context

Change-Id: I45f4f5951fedd663616e2275b047dcc02ed77a74