fix(tests): fix client eslint rules also applying to tests
Also, this adds the wikimedia config for jest. If we already have it, we
might just as well use it, too.

The browser env is necessary so that eslint knows about `document` being
globally available (which it is because we use jsdom).

Finally, jest.config.js has fallen through the cracks in terms of
linting jest config files.

Change-Id: I8749534523a90a0d3576a0d05e0ab97b7c0fc7ef