feat(editor): Add min/max-length validation to string control
jsonrainbow seems to be using either `mb_strlen` or `iconv_strlen` to
calculate the length of the string for validation. Thus we cannot use
the native HTML-attributes for that or the naive `'abc'.length` value in
javascript, but need an approach that actually mimics what our PHP
library is doing.

Also, this change trims the input, but for validation as well as for
submission.

This change uses the same copy as the jsonrainbow library, though that
can of course be adjusted.

Bug: T372850
Bug: T373160
Change-Id: Id440aa8a1f704ccdd27b174a5f8f87bd047b8491