refactor(editor): move `step` attr from useJsonFormControl to NumberControl
That's a number-specific attribute that should not live in a generic
useJsonFormControl method. Further, this will become slightly more
complex as we add support for multipleOf in NumberControl as well.

Finally, this extracts a bunch of bootstrapping methods to consolidate
the boilerplate of the test and extracts a variable in the control for
ease of reading.

Change-Id: I9619ec0600ec5d9e5857aafdffe9a83b689181f7