chore: add trailing commas in all PHP files
Briefly, trailing commas make it easier to move elements around in an
object or array, makes it easier to add/delete elements at the end and
results in cleaner diffs.

This is step 1: disabling the rule and adding all the commas. In step 2,
we will re-enable the rule with the opposite effect: requiring trailing
commas. The goal here is to make it easier to revert changes that were
added when we were still forbidding trailing commas.

Bug: T364771
Change-Id: I19355c8cb81e19283ddc9c6f1f5a079412d3d110