refactor(MediaWikiConfigReader): inject LoggerInterface
In general, injecting the logger via a constructor parameter should be
our default approach over making classes LoggerAware (either via
interface directly or using trait). See also the performance review in
T361070#9758085.

In this specific case, we also do not want other consumers to be able
change the logger, because as services are singletons, this would change
the logger for _all_ consumers of that service, not just the one that
sets a different logger.

Change-Id: Ic420ecd02469961c3ef8e74ce2afb13dc7b5d5a3