tests: Improve coverage
The goal of the patch is to improve coverage
for ConfigurationProviderFactory. To do that,
it switches to DI for ExtensionRegistry and
ConfigurationProviderFactoryTest, which allows
us to mock the registry (this is necessary to avoid
installed extensions registering their own data).

This also allows mocking ExtensionRegistry for Utils,
so a test for that class is added as well.

Change-Id: I804f87c6b8817fc7d29e535f68fdbe7d7406a817