refactor: Split path-related logic from EmergencyDefaultsUpdater
I am about to write logic to read emergency defaults into memory.
For that, I need to know where to find relevant files to require.
Currently, building up relevant paths is in the updater. However,
it does not make sense to construct the updater for reader.
This patch creates an independent service to construct the paths,
which would be used in both updater and reader.

Keeping old methods as deprecated, as they are now in use
from other extension's tests.

Bug: T367516
Change-Id: Icf39d199c2a9d458a6df41a0b0de08bdd4b80d65