fix(CommonsFileControl): don't load full-sized images as menu thumbnails
Without this change, we load the images at their original resolutions
from commons in order to display them as tiny 40x40 pixel thumbnails.
That uses a lot of bandwith and is thus very slow.

Also, while being here, it removes the `showThumbnail` property which
has no effect in the MenuItemData type, only in the MenuConfig type
where we already have set it statically to `true`.

As far as I can tell, the `width` and `height` values in the Thumbnail
type are actually unused by Codex, but I inquired about them upstream:
https://wikimedia.slack.com/archives/C03DKGSEPL2/p1729853977120869
For now, the value of the width is set to 40px which is both the
requested width and the width that is displayed in the menu, as far as I
can tell.

80px is chosen as the image width to ensure a crisp thumbnail also on
high-resolution screens.

Change-Id: Ia3310018c0ab837c358bbe17b45bee7bb13031cc