[i18n] Drop unused communityconfiguration-unknown-error
The message was introduced in 332359a00755c4d271a0e41dfd16af5a9aa64711,
where it was used in SpecialCommunityConfiguration. This usage was
dropped in cc8425c788379395a2a0431474321d78ab143e09, where editor
capabilities were introduced. However, the message itself was not
dropped.

According to git grep and codesearch, the message is currently not
in use. As such, let's remove it from the codebase.

Change-Id: Ib1323ee33a70155ec96b216cb481047bb33d718a