MediaWikiConfigReader: Log when falling back
If MediaWikiConfigReader is used to load a configuration
variable that is not registered by any community configuration
provider, it is a sign of an issue.

I discovered GrowthExperiments does this with GECampaigns
and GECampaignTopics (uses community configuration to get
those, but does not register them in any schema), which happened
because of historical reasons (legacy CC worked on different principles,
and this was used to make something stored on-wiki without actually
bothering the community too much with it).

I would like to know what other variables are used in a similar
manner.

Bug: T384184
Change-Id: I4157f0ca30aa6024af58bb56aa4bd8e0a6498172