Show subpages of Special:CommunityConfiguration in search
Special pages can define the getSubpagesForPrefixSearch()
method, which then informs Search what subpages are available.
Make use of that functionality to return the list of providers
available. This makes it simpler to link to a given CC form,
as opposed to linking the full dashboard. It also makes navigation
using the searchbox simpler for users.

Bug: T362043
Change-Id: Iacbd9b48b2577e63ce7a499c02f52ace6b918d7e